pyproject.toml
[project]
name = "task-management-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "anyio>=4.10.0",
    "asyncpg>=0.30.0",
    "email-validator>=2.3.0",
    "faker>=37.8.0",
    "fastapi>=0.117.1",
    "fastapi-cache2>=0.2.2",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "pwdlib[argon2]>=0.2.1",
    "pydantic-settings>=2.10.1",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "redis>=7.0.0",
    "sqlalchemy[asyncio]>=2.0.43",
    "sqlmodel>=0.0.25",
    "uvicorn>=0.36.0",
]
